refactor : SRP 적용 코드 리펙토링
