+++ src/main/java/com/example/concertreservationsystem/application/scheduler/TokenActivateScheduler.java
@@ -1,6 +1,8 @@
 package com.example.concertreservationsystem.application.scheduler;
 
 
+import com.example.concertreservationsystem.domain.service.queue.QueueService;
+import com.example.concertreservationsystem.domain.service.queue.TokenService;
 import com.example.concertreservationsystem.domain.service.reservation.ReservationService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -14,14 +16,14 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class TokenActivateScheduler {
 
-    private final ReservationService reservationService;
+    private final TokenService tokenService;
 
     // 스케줄러 실행
     @Transactional
     @Scheduled(fixedRate = 60000)  // 60초마다 작업 수행시간과 상관없이 실행
     public void cancelExpiredReservations() {
         log.info("======== active_tokens 활성화 스케쥴러 실행 시작 ================");
-        reservationService.activateTokens();
+        tokenService.activateTokens();
         log.info("========  active_tokens 활성화 스케쥴러 실행완료 ================");
     }
 }
