+++ src/main/java/com/example/concertreservationsystem/domain/service/concert/ConcertService.java
@@ -1,9 +1,12 @@
 package com.example.concertreservationsystem.domain.service.concert;
 
 import com.example.concertreservationsystem.domain.model.Concert;
+import com.example.concertreservationsystem.domain.model.Seat;
 import com.example.concertreservationsystem.domain.repo.ConcertRepository;
 import com.example.concertreservationsystem.application.concert.dto.request.ConcertRequestDto;
 import com.example.concertreservationsystem.application.concert.dto.response.ConcertResponseDto;
+import com.example.concertreservationsystem.infrastructure.persistence.JpaConcertRepository;
+import com.example.concertreservationsystem.infrastructure.persistence.JpaSeatRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -32,4 +35,10 @@ public class ConcertService{
 
         return new ConcertResponseDto(concert.getName(), concert.getPrice());
     }
+
+    public Concert getConcertById(Long concertId) {
+        return concertRepository.findById(concertId)
+                .orElseThrow(() -> new IllegalArgumentException("해당하는 콘서트가 없습니다."));
+    }
+
 }
