+++ src/main/java/com/example/concertreservationsystem/domain/service/user/UserService.java
@@ -8,16 +8,13 @@ import com.example.concertreservationsystem.application.user.dto.request.UserReq
 import com.example.concertreservationsystem.application.user.dto.response.UserPointResponseDto;
 import com.example.concertreservationsystem.application.user.dto.response.UserPositionResponseDto;
 import com.example.concertreservationsystem.application.user.dto.response.UserResponseDto;
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.PersistenceContext;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
-import java.util.List;
+
 
 @Slf4j
 @Service
@@ -27,9 +24,6 @@ public class UserService  {
     private final UserRepository userRepository;
     private final ReservationService reservationService;
     private final RedisTemplate<String,Object> redisTemplate;
-
-//    @PersistenceContext
-//    private EntityManager entityManager;
 
 
     @Transactional
@@ -55,7 +49,7 @@ public class UserService  {
     public UserPointResponseDto chargePoint(String token, UserPointRequestDto requestDto) {
 
         // 대기열 토큰 검증
-        User user = reservationService.validateAnyToken(token);
+        User user = validateAnyToken(token);
 
         // 잔액 충전
         user.addPoints(requestDto.getPoint());
@@ -66,14 +60,14 @@ public class UserService  {
 
     // 잔액 조회 (대기열 토큰 필수)
     public UserPointResponseDto getUserPoint(String token) {
-        User user = reservationService.validateAnyToken(token);
+        User user = validateAnyToken(token);
         return new UserPointResponseDto(user.getPoint());
     }
 
     // 본인의 대기번호 조회
     public UserPositionResponseDto getUserQueuePosition(String token) {
         // Redis를 이용한 대기열 토큰 검증
-        reservationService.validateAnyToken(token);
+        validateAnyToken(token);
 
         // 토큰의 대기열 순번 조회
         Long position = redisTemplate.opsForZSet().rank("waiting_queue", token);
@@ -89,35 +83,57 @@ public class UserService  {
         return new UserPositionResponseDto(position);
     }
 
+    // 대기열 토큰 여부 검증
+    // 대기열 토큰 검증 메서드
+    public User validateToken(String queueToken) {
+        // 활성화된 토큰인지 확인
+        if (!redisTemplate.opsForSet().isMember("active_tokens", queueToken)) {
+            throw new IllegalArgumentException("유효하지 않은 토큰입니다.");
+        }
+        // 토큰에 연결된 유저 정보 조회
+        Object userId = redisTemplate.opsForHash().get("queue:token:" + queueToken, "userId");
+        if (userId == null) {
+            throw new IllegalArgumentException("토큰에 해당하는 유저 정보를 찾을 수 없습니다.");
+        }
+        String userIdStr = userId.toString();
+        Long checkuserId;
 
-//    @Transactional
-//    public void generateDummyUsers(int count) {
-//        List<User> userList = new ArrayList<>();
-//        int batchSize = 1000;
-//
-//        for (int i = 1; i <= count; i++) {
-//            String username = "User" + i;
-//            User user = User.builder()
-//                    .name(username)
-//                    .build();
-//            userList.add(user);
-//
-//            if (i % batchSize == 0) {
-//                userRepository.saveAll(userList);
-//                userRepository.flush();
-//                userList.clear();
-//                entityManager.clear();
-//            }
-//        }
-//
-//        if (!userList.isEmpty()) {
-//            userRepository.saveAll(userList);
-//            userRepository.flush();
-//            userList.clear();
-//            entityManager.clear();
-//        }
-//    }
-//    public long countUsers() {
-//        return userRepository.count();
-//    }
+        try {
+            checkuserId = Long.parseLong(userIdStr);
+        } catch (NumberFormatException e) {
+            throw new IllegalArgumentException("유효하지 않은 ID 형식");
+        }
+
+        return userRepository.findById(checkuserId)
+                .orElseThrow(() -> new IllegalArgumentException("유저 정보를 찾을 수 없습니다."));
+    }
+
+    public User validateAnyToken(String token) {
+        // 1. 활성화된 토큰인지 확인
+        Boolean isActive = redisTemplate.opsForSet().isMember("active_tokens", token);
+
+        // 2. 대기열에 있는 토큰인지 확인
+        Long rank = redisTemplate.opsForZSet().rank("waiting_queue", token);
+
+        if (Boolean.TRUE.equals(isActive) || rank != null) {
+            // 토큰에 연결된 유저 정보 조회
+            Object userId = redisTemplate.opsForHash().get("queue:token:" + token, "userId");
+            if (userId == null) {
+                throw new IllegalArgumentException("토큰에 해당하는 유저 정보를 찾을 수 없습니다.");
+            }
+            String userIdStr = userId.toString();
+            Long checkUserId;
+
+            try {
+                checkUserId = Long.parseLong(userIdStr);
+            } catch (NumberFormatException e) {
+                throw new IllegalArgumentException("유효하지 않은 ID 형식입니다.");
+            }
+
+            return userRepository.findById(checkUserId)
+                    .orElseThrow(() -> new IllegalArgumentException("유저 정보를 찾을 수 없습니다."));
+        } else {
+            throw new IllegalArgumentException("유효하지 않은 토큰입니다.");
+        }
+    }
 }
