+++ src/test/java/com/example/concertreservationsystem/event/EventUnitTest.java
@@ -3,8 +3,8 @@ package com.example.concertreservationsystem.event;
 import com.example.concertreservationsystem.domain.model.Concert;
 import com.example.concertreservationsystem.domain.model.ConcertEvent;
 import com.example.concertreservationsystem.domain.repo.ConcertRepository;
+import com.example.concertreservationsystem.domain.service.concert.SeatInitializer;
 import com.example.concertreservationsystem.domain.service.event.ConcertEventService;
-import com.example.concertreservationsystem.infrastructure.config.SeatInitializer;
 import com.example.concertreservationsystem.infrastructure.persistence.JpaConcertEventRepository;
 import com.example.concertreservationsystem.application.event.dto.request.EventRequestDto;
 import com.example.concertreservationsystem.application.event.dto.response.EventResponseDto;
